11.1.0

+++ UpdateFieldCodeGenerator/Structures/JamMirrorCTROptions_C.cs
@@ -2,8 +2,8 @@
 {
     public class JamMirrorCTROptions_C
     {
-        public static readonly UpdateField m_conditionalFlags = new UpdateField(typeof(int), UpdateFieldFlag.None);
-        public static readonly UpdateField m_factionGroup = new UpdateField(typeof(uint), UpdateFieldFlag.None);
+        public static readonly UpdateField m_conditionalFlags = new UpdateField(typeof(uint), UpdateFieldFlag.None);
+        public static readonly UpdateField m_factionGroup = new UpdateField(typeof(byte), UpdateFieldFlag.None);
         public static readonly UpdateField m_chromieTimeExpansionMask = new UpdateField(typeof(uint), UpdateFieldFlag.None);
     }
 }
